# Fix race condition in Telemetry and improve throughput