@qqqlab 2023-02-23
Hi Jye, thanks for your feedback. I'm not an expert on the git administrative side of ELRS, but would qualify this both as bug fix and maintenance branch.

My main goal is to get bi-directional telemetry working smoothly and with maximum throughput in both directions. The bug fix alone made some improvement, but adding the double buffer strategy really did the trick.

The bug fix alone would be: replace `if (targetFound)` with `if (targetFound && !payloadTypes[targetIndex].locked)` at line 301.