@qqqlab 2023-03-01
Sorry, I do not agree. The datalink has limited throughput, if the FC sends more info than can be handled by the link then data will be dropped regardless of how many slots or queue positions are available. Adding a longer queue might even work in the opposite direction as it increases latency. It would more sense to me to add a flow control mechanism instead, i.e. to inform the FC about the queue state.

In the current design no specific slots are assigned for specific extended CRSF messages. The two available slots are both used for multiple types of messages. 

One could argue for adding a separate slot for "status text" in the new design. I did not do it, because I think this is sufficiently handled by the queue.

Yes, one test is failing now, but that is a direct consequence of the proposed design change. It is intended behavior.