@schugabe 2023-03-01
> Sorry, I do not agree. The datalink has limited throughput, if the FC sends more info than can be handled by the link then data will be dropped regardless of how many slots or queue positions are available. Adding a longer queue might even work in the opposite direction as it increases latency. It would more sense to me to add a flow control mechanism instead, i.e. to inform the FC about the queue state.

We don't have a "slow down" crsf message type so it's not possible to implement that right now but we are trying to work out a way crsf can be extended: https://github.com/tbs-fpv/freedomtx/issues/26 

> In the current design no specific slots are assigned for specific extended CRSF messages. The two available slots are both used for multiple types of messages.
> 
> One could argue for adding a separate slot for "status text" in the new design. I did not do it, because I think this is sufficiently handled by the queue.
> Yes, one test is failing now, but that is a direct consequence of the proposed design change. It is intended behavior.

Merging code that breaks working features is not an option for me and if calling it a design descision does not change that.

My suggestion: create a new PR that fixes the race condition. The remaining changes need more discussion and/or new crsf frames.