@qqqlab 2023-03-02
Thanks for your feedback. 

@schugabe Please, let's discuss this further. 

My initial thought was to keep the existing code as-is, and just add a fifo buffer to improve throughput. But the existing code for extended crsf did not really make sense to me. For example, the code mentions "reserve slot for status text", but the same slot is also used for msp and any non explicitly handled crsf packets.

Also, I asked myself the rather philosophical question: should elrs keep an ever increasing list of crsf messages and dictate priorities for them, or should elrs be net-neutral, treat all messages equally, and just provide the best possible link to the communicating parties? 

I like the second option better.

Based on this I submitted a more radical PR, making sure that the PR still satisfies the requirements for extended crsf as outlined in the existing code. .

If it is a hard requirement that all current test pass, I can add code back in to make that happen. But before I spend time on that I would like to get feedback from you if there is an interest in proceeding with this PR.