@wvarty 2023-03-04
@qqqlab just an FYI, we are aiming to cut a minor bugfix release towards the end of March, and hope to have all fixes merged to the 3.x.x-maintenance branch before the 9th March to leave time for running the test plan.
It might be worth splitting out the bugfix in this PR to a new PR that targets 3.x.x-maintenance, and leave the enhancements here for discussion once @schugabe gets a chance to respond?