@qqqlab 2023-04-29
@schugabe 

> Basically we are back to the old logic and need 64 bytes more buffer space only for ardupilot text messages.

Thanks for your feedback. Apparently I fail to clearly explain what this PR is about. This is not about the special cases (standard CRSF header GPS, SPEED, etc., or extended header CRSF status text etc.) but about any other CRSF message. 

With the special cases we're interested in the most recent message, so it is fine to use a single buffer, so that we receive the most recent message and potentially drop the ones that came before. But with the rest we're interested to get each and every message. By using a single buffer for these messages, only 50% of the OTA capacity is used in worse case. Using a double buffer fixes this deficiency.

See example below: a CRSF message arrives every 4 transmission slots, and takes 5 slots to transmit OTA. With single buffering only 3 out of 6 messages are received OTA. With double buffering 4.8 of 6 messages are received, because while a message is being transmitted OTA the next incoming message is buffered in the second buffer.
```
incoming:   1---2---3---4---5---6---  
OTA single: 11111---33333---55555---
OTA double: 111112222233333444446666
V3.2.0:     11112---33334---55556---

('-' is idle transmission slot, '1' to '6' is msg transmission slot)
```

With single buffering, when the incoming data rate increases from 100% to 101% of OTA link capacity, OTA data rate suddenly drops from 100% to 50% and 51% of data is lost. This makes things difficult to diagnose. With double buffering the situation is more predictable, the OTA rate stays fully loaded at 100% and only 1% of incoming data is lost. 

Note that it is very easy to overload the link, incoming data flows at 400kbps, whereas the OTA capacity is less than 16kbps.

Regarding current use cases:

V3.2.1 includes PR https://github.com/ExpressLRS/ExpressLRS/pull/2103 which fixes a race condition. For illustrative purposes the situation with V3.2.0 is also depicted above, now 0 out of 6 messages are received because the messages get corrupted by the next incoming message which overwrites the single buffer as it is being transmitted OTA.

Prior to V3.2.1 all extended CRSF messages (including all Ardupilot messages) were suffering from this race condition. Apparently everybody (including myself) was happy with the situation, as apparently nobody complained about missing messages. I only noticed the deficiency when I tried to send more data over the link. 

Maybe others also noticed this, but instead spending the time to investigate, they switched to another datalink solution.

So, it appears to me that in the current use cases the CRSF messages are sufficiently throttled by the FC not to trigger the race condition. This also means that switching to double buffering will have no negative impact on these use cases.

Regarding the StatusText special case: 

If an additional 64 bytes of RAM is unacceptable for low-RAM targets, we can either reverse the last commit, or use it only for plenty-RAM ESP targets.