@AlessandroAU 2023-05-09
@qqqlab We have a very active private developer channel on our discord server, It would be great for you to join and discuss ideas in real-time. Come join and tag Jye, Wez, PK or myself and we will add you.